+++ standingssync/__init__.py
@@ -1,3 +1,4 @@
 default_app_config = 'standingssync.apps.SyncAltContactsConfig'
 
-__version__ = '1.0.0'
+__version__ = '1.0.1'
+__title__ = 'Standings Sync'