+++ CHANGELOG.md
@@ -13,6 +13,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [1.0.1] - 2019-12-19
+
+### Changed
+
+- Improved error handling of contacts sync process
+- Improved layout of messages
+
 ## [1.0.0] - 2019-10-02
 
 ### Added